move norm message submission to NormLayout
